requirements.txt
python-dotenv
pyupbit
pyjwt
pandas
pandas_ta
schedule